requirements.txt
# --- Runtime core ------------------------------------------------------------
alembic==1.16.5
fastapi==0.116.1
starlette==0.47.3            # aliniat cu FastAPI 0.116.x
httpx[http2]==0.27.2         # necesar pentru Emag SDK; include h2
h2==4.1.0                    # pin explicit pt. http2 stabil
anyio==4.10.0                # compat cu starlette 0.47.x
typing-extensions==4.15.0

SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-core==2.33.2        # pin pt. reproducibilitate
tenacity==9.1.2
requests==2.32.3
Mako==1.3.10
python-dotenv==1.0.1
uvicorn[standard]==0.30.6    # include uvloop, httptools, websockets, watchfiles

# --- (opțional) Performanță / QoL -------------------------------------------
# orjson==3.10.7             # FastAPI îl folosește automat dacă e prezent (JSON mai rapid)
# python-multipart==0.0.9    # necesar DOAR dacă expui upload de fișiere

# --- Teste (poți muta într-un requirements-dev.txt dacă vrei imagini mai mici) ---
pytest==8.4.1
pytest-asyncio==0.23.8     # util dacă adaugi teste async
cachetools>=5.3
respx>=0.21

psycopg[binary]==3.2.1